Moved api to different class file
